TANDEM Edge/TANDEM Edge Orchestrator/location-apis/alert_app/requirements.txt
influxdb==5.2.2
python-dateutil==2.8.0
paho-mqtt==1.4.0